refactor: move retry_request into a shared lib
