+++ backend/app/modules/multisig_signatures/service/offchain.py
@@ -6,6 +6,8 @@ from app.extensions import db
 from app.infrastructure import database
 from app.infrastructure.database.models import MultisigSignatures
 from app.infrastructure.database.multisig_signature import SigStatus, MultisigFilters
+
+from app.infrastructure.external_api.common import retry_request
 from app.modules.common.allocation_deserializer import deserialize_payload
 from app.modules.common.crypto.eip1271 import get_message_hash
 from app.modules.common.crypto.signature import (
@@ -21,10 +23,7 @@ from app.modules.multisig_signatures.core import (
 )
 from app.modules.multisig_signatures.dto import Signature
 from app.pydantic import Model
-from app.infrastructure.external_api.safe.message_details import (
-    get_message_details,
-    retry_request,
-)
+from app.infrastructure.external_api.safe.message_details import get_message_details
 
 
 class OffchainMultisigSignatures(Model):
